+++ fsleyes/plugins/__init__.py
@@ -8,8 +8,9 @@
 
 
 FSLeyes uses a simple plugin architecture for loading custom views, controls,
-and tools. Plugins can be installed from Python libraries (e.g. as hosted on
-`PyPi <https://pypi.org/>`_), or installed directly from a ``.py`` file.
+tools, and layouts. Plugins can be installed from Python libraries (e.g. as
+hosted on `PyPi <https://pypi.org/>`_), or installed directly from a ``.py``
+file.
 
 
 In both cases, FSLeyes uses `entry points
@@ -21,7 +22,7 @@ Things plugins can provide
 --------------------------
 
 
-FSLeyes plugins can provide custom *views*, *controls* and *tools*:
+FSLeyes plugins can provide custom *views*, *controls*, *tools*, and *layouts*:
 
  - A *view* is a top level panel, such as an :class:`.OrthoPanel`,
    :class:`.Scene3DPanel`, or :class:`.TimeSeriesPanel`. Views provided
@@ -36,6 +37,29 @@ FSLeyes plugins can provide custom *views*, *controls* and *tools*:
    the top-level *Tools* menu, such as the :class:`.ApplyFlirtXfmAction`, the
    :class:`.CropImageAction`, and the :class:`.ResampleAction`.
 
+ - A *layout* is a string which specifies the layout of the FSLeyes frame,
+   comprising one or more view panels, and a set of control panels for each
+   view. Refer to the :mod:`fsleyes.layouts` module for more details.
+
+
+FSLeyes plugin sources
+----------------------
+
+FSLeyes plugins can be loaded from the following locations:
+
+ - Built-in plugins from the :mod:`fsleyes.plugins` package.
+ - Single-file plugins that have been loaded/installed by the user.
+ - Plugins from third-party libraries that are installed into the running Python
+   environment.
+
+The default behaviour, when FSLeyes starts up, is to only expose plugins from
+the first two locations - plugins from third party libraries are hidden by
+default. However, third-party plugins are automatically made available when a
+layout from the same library is loaded.
+
+Third-party plugins can also be made visible by default if you start FSLeyes
+with the ``--showAllPlugins`` command-line option.
+
 
 Loading/installing FSLeyes plugins
 ----------------------------------
@@ -47,8 +71,8 @@ FSLeyes plugins are loaded into a running FSLeyes as follows:
    environment that FSLeyes is running in, and which provide any FSLeyes entry
    points, will automatically be detected by FSLeyes.
 
- - Plugin ``.py`` files, which contain view, control, and/or tool definitions,
-   can be passed directly to the :func:`loadPlugin` function.
+ - Plugin ``.py`` files, which contain view, control, tool, and/or layout
+   definitions, can be passed directly to the :func:`loadPlugin` function.
 
  - Plugin ``.py`` files which are present in the FSLeyes settings directory,
    or which are found in the ``FSLEYES_PLUGIN_PATH`` environment variable, will
@@ -77,13 +101,14 @@ Writing a FSLeyes plugin
           built-in plugins can be found in ``fsleyes/plugins/``.
 
 
-.. warning:: FSLeyes assumes that all views, controls, and tools have unique
-             class names.  So expect problems if, for example, you define your
-             own FSLeyes control with the name ``OverlayListPanel``.
+.. warning:: FSLeyes assumes that **all** views, controls, tools, and layouts
+             have unique names.  So expect problems if, for example, you
+             define your own FSLeyes control with a name that is already used
+             by a built-in control, e.g. ``OverlayListPanel``.
 
 
 A FSLeyes plugin is a Python library, or a ``.py`` file, which contains
-definitions for custom views, controls, and tools.
+definitions for custom views, controls, tools, and layouts.
 
  - Views must be sub-classes of the :class:`.ViewPanel` class.
 
@@ -92,10 +117,16 @@ definitions for custom views, controls, and tools.
 
  - Tools must be sub-classes of the :class:`.Action` class.
 
+ - Layouts must be strings or tuples. For single-file plugins, layout variables
+   must have a name that begins with ``FSLEYES_LAYOUT_``. If a layout is a
+   string, it is given a name based on the name of the variable. If a layout
+   is a tuple, the first value in the tuple is used as the name, and the second
+   value is assumed to be the layout string.
 
-To write a ``.py`` file which can be loaded as a FSLeyes plugin, simply
-define your views, controls, and tools in the file. The file path can then
-be passed to the :func:`loadPlugin` or :func:`installPlugin` function.
+
+To write a ``.py`` file which can be loaded as a FSLeyes plugin, simply define
+your views, controls, tools, and layouts in the file. The file path can then be
+passed to the :func:`loadPlugin` or :func:`installPlugin` function.
 
 
 To release a FSLeyes plugin as a library, you need to organise your code
@@ -103,7 +134,7 @@ as a Python library. Minimally, this requires the following:
 
  - Arrange your ``.py`` file(s) into a Python package.
 
- - Expose your custom views, controls, and tools as `entry points
+ - Expose your custom views, controls, tools, and layouts as `entry points
    <https://packaging.python.org/specifications/entry-points/>`_.
 
 
@@ -116,10 +147,11 @@ A minimal ``pyproject.toml`` file for a FSLeyes plugin might look like this::
     [project]
     name = "my-cool-fsleyes-plugin"
 
-    # Views, controls, and tools must be exposed
-    # as entry points within groups called
-    # "fsleyes_views", "fsleyes_controls" and
-    # "fsleyes_tools" respectively.
+    # Views, controls, tools, and layouts
+    # must be exposed as entry points
+    # within groups called "fsleyes_views",
+    # "fsleyes_controls", "fsleyes_tools"
+    # and "fsleyes_layouts" respectively.
 
     [project.entry-points.fsleyes_views]
     "My cool view" = "myplugin:MyView"
@@ -129,6 +161,9 @@ A minimal ``pyproject.toml`` file for a FSLeyes plugin might look like this::
 
     [project.entry-points.fsleyes_tools]
     "My cool tool" = "myplugin.MyTool"
+
+    [project.entry-points.fsleyes_layouts]
+    "My cool layout" = "myplugin.MyLayout"
 
 
 See the `Python Packaging guide
@@ -172,17 +207,22 @@ The following functions can be used to access plugins:
    listViews
    listControls
    listTools
-   lookupView
+   listLayouts
    lookupControl
    lookupTool
+   layoutModule
    pluginTitle
 """
 
 
+import functools           as ft
 import os.path             as op
 import                        os
 import                        sys
 import                        glob
+import                        random
+import                        string
+import                        fnmatch
 import                        pkgutil
 import                        logging
 import                        importlib
@@ -190,7 +230,8 @@ import importlib.util      as imputil
 import importlib.metadata  as impmeta
 import                        collections
 
-from typing import Dict, Union, Type, Optional, Iterator
+
+from typing import Dict, Union, Type, Optional, Sequence
 from types  import ModuleType
 
 import fsl.utils.settings            as fslsettings
@@ -204,20 +245,45 @@ import fsleyes.controls.controlpanel as ctrlpanel
 log = logging.getLogger(__name__)
 
 View    = Type[viewpanel.ViewPanel]
-Control = Type[ctrlpanel.ControlPanel]
+Control = Union[Type[ctrlpanel.ControlMixin], Type[ctrlpanel.ControlToolBar]]
 Tool    = Type[actions.Action]
-Plugin  = Union[View, Control, Tool]
+Layout  = Union[str, tuple]
+Plugin  = Union[View, Control, Tool, Layout]
 
 
-def class_defines_method(cls : Type, methname : str) -> bool:
-    """Check to see whether ``methname`` is implemented on ``cls``, and not
-    on a base-class.
+SHOW_THIRD_PARTY_PLUGINS = set()
+"""Global toggle which controls whether plugins provided by installed
+third-party libraries are exposed by the :func:`listViews`,
+:func:`listControls`, and :func:`listTools` functions.  Layouts provided
+by plugins are always visible.
 
-    :meth:`.Action.ignoreTool`, :meth:`.ControlMixin.ignoreControl`, and
-    :meth:`.ControlMixin.supportSubClasses` need to be implemented on the
-    specific class - inherited base class implementations are not considered.
+This field may either be a set containing the names of specific third-party
+plugins to show, or a boolean which will toggle all third party plugins on or
+off.
+"""
+
+
+def showThirdPartyPlugin(modname : str):
+    """Show plugins from the given third party module. """
+    global SHOW_THIRD_PARTY_PLUGINS
+    if SHOW_THIRD_PARTY_PLUGINS is True:
+        return
+    if SHOW_THIRD_PARTY_PLUGINS is False:
+        SHOW_THIRD_PARTY_PLUGINS = set()
+    SHOW_THIRD_PARTY_PLUGINS.add(modname)
+
+
+def shouldShowThirdPartyPlugin(modname : str) -> bool:
+    """Return ``True`` if the given plugin should be visible, ``False``
+    otherwise.
     """
-    return methname in cls.__dict__
+    if SHOW_THIRD_PARTY_PLUGINS in (True, False):
+        return SHOW_THIRD_PARTY_PLUGINS
+
+    for pattern in SHOW_THIRD_PARTY_PLUGINS:
+        if fnmatch.fnmatch(modname, f'{pattern}*'):
+            return True
+    return False
 
 
 class FSLeyesPlugin(impmeta.Distribution):
@@ -228,21 +294,16 @@ class FSLeyesPlugin(impmeta.Distribution):
     is registered with the :meth:`FSLeyesPluginFinder.add_plugin` method.
     """
 
-    def __init__(self,
-                 module  : ModuleType,
-                 modname : str,
-                 builtin : bool):
+    def __init__(self, module : ModuleType, modname : str):
         """Create a ``FSLeyesPlugin`` from a single-file plugin file that
         has already been loaded as a module.
 
         :arg module:  The loaded module
         :arg modname: The module name
-        :arg builtin: Whether or not this is a built-in FSLeyes plugin, from
-                      the :mod:`fsleyes.plugins` module.
         """
+        log.debug('New FSLeyesPlugin(%s)', modname)
         self.__module  = module
         self.__modname = modname
-        self.__builtin = builtin
 
 
     @property
@@ -256,58 +317,72 @@ class FSLeyesPlugin(impmeta.Distribution):
 
 
     @property
-    def entry_points(self) -> Iterator[impmeta.EntryPoint]:
+    @ft.lru_cache
+    def entry_points(self) -> Sequence[impmeta.EntryPoint]:
         """Return a sequence of ``EntryPoint`` objects provided by the plugin.  The
         :meth:`FSLeyesPlugin.find_entry_points` function is used to scan the
         module for entry points.
         """
 
         modname = self.__modname
-        alleps  = FSLeyesPlugin.find_entry_points(
-            self.__module, not self.__builtin)
+        alleps  = FSLeyesPlugin.find_entry_points(self.__module)
+        epobjs  = []
 
         for group, eps in alleps.items():
-            for name, cls in eps.items():
+            for name, plugin in eps.items():
 
-                label = cls.title()
-                # Look up label for built-in plugins
-                if label is None:
-                    if group == 'fsleyes_tools':
-                        label = strings.actions.get(name, name)
+                # layouts can be either strings with
+                # variable name FSLEYES_LAYOUT_<label>
+                # or tuples containing (label, layout)
+                if isinstance(plugin, (tuple, str)):
+                    if isinstance(plugin, tuple):
+                        label  = plugin[0]
+                        plugin = plugin[1]
                     else:
-                        label = strings.titles.get(name, name)
+                        label = name[15:]
 
-                yield impmeta.EntryPoint(label,
-                                         f'{modname}:{name}',
-                                         group=group)
+                # View/Control/Tool plugins can implement
+                # a title() staticmethod to specify their
+                # title/label
+                else:
+                    label = plugin.title()
+                    # Look up label for built-in plugins
+                    if label is None:
+                        if group == 'fsleyes_tools':
+                            label = strings.actions.get(name, name)
+                        else:
+                            label = strings.titles.get(name, name)
+
+                epobjs.append(impmeta.EntryPoint(label,
+                                                 f'{modname}:{name}',
+                                                 group=group))
+        return epobjs
 
 
     @staticmethod
-    def find_entry_points(
-            mod            : ModuleType,
-            ignoreBuiltins : bool
-    ) -> Dict[str, Dict[str, Plugin]]:
+    def find_entry_points(mod : ModuleType) -> Dict[str, Dict[str, Plugin]]:
         """Finds the FSLeyes entry points (views,
-        controls, or tools) that are defined within the given module.
+        controls, tools, or layouts) that are defined within the given module.
 
-        :arg mod:            The module to search
-        :arg ignoreBuiltins: If ``True``, all, views, controls and tools which
-                             are built into FSLeyes will be ignored. The
-                             :class:`.ViewPanel`, :class:`.ControlPanel`,
-                             :class:`.ControlToolBar` and :class:`.Action` base
-                             classes are always ignored.
-
-        :returns:            A dictionary
+        :arg mod: The module to search
+        :returns: A dictionary
         """
 
         entryPoints = collections.defaultdict(dict)
 
         for name in dir(mod):
 
-            item = getattr(mod, name)
+            item  = getattr(mod, name)
+            ptype = _pluginType(item)
 
-            if not isinstance(item, type):
+            if not ptype:
                 continue
+
+            if ptype == 'layout':
+                # Layouts from single-file plugins must have
+                # a variable name starting with FSLEYES_LAYOUT_
+                if not name.startswith('FSLEYES_LAYOUT_'):
+                    continue
 
             bases = [viewpanel.ViewPanel,
                      canvaspanel.CanvasPanel,
@@ -316,22 +391,32 @@ class FSLeyesPlugin(impmeta.Distribution):
                      ctrlpanel.SettingsPanel,
                      actions.Action]
 
-            # avoid base-classes and built-ins
+            # avoid base-classes
             if item in bases:
                 continue
-            if ignoreBuiltins and str(item.__module__).startswith('fsleyes.'):
-                continue
 
-            # ignoreControl/ignoreTool may be overridden
-            # to tell us to ignore this plugin
-            if issubclass(item, ctrlpanel.ControlMixin)    and \
-               class_defines_method(item, 'ignoreControl') and \
-               item.ignoreControl():
-                continue
-            if issubclass(item, actions.Action)         and \
-               class_defines_method(item, 'ignoreTool') and \
-               item.ignoreTool():
-                continue
+            # ignore imported classes - e.g. a module
+            # might import an existing plugin like so,
+            # for the sake of being sub-classed - in
+            # this case, we only want to consider
+            # MyView:
+            #     from blah import SomeView:
+            #     class MyView(SomeView):
+            #         ..
+            if ptype in ('view', 'control', 'tool'):
+                if item.__module__ != mod.__name__:
+                    continue
+
+                # ignoreControl/ignoreTool may be overridden
+                # to tell us to ignore this plugin
+                if ptype == 'control'               and \
+                   'ignoreControl' in item.__dict__ and \
+                   item.ignoreControl():
+                    continue
+                if ptype == 'tool'               and \
+                   'ignoreTool' in item.__dict__ and \
+                   item.ignoreTool():
+                    continue
 
             group = _pluginGroup(item)
             if group is not None:
@@ -351,9 +436,9 @@ class FSLeyesPluginFinder(importlib.abc.MetaPathFinder):
 
         instance = getattr(FSLeyesPluginFinder, '_instance', None)
         if instance is None:
-            instance = FSLeyesPluginFinder()
+            instance                      = FSLeyesPluginFinder()
+            FSLeyesPluginFinder._instance = instance
             sys.meta_path.append(instance)
-
         return instance
 
 
@@ -361,19 +446,16 @@ class FSLeyesPluginFinder(importlib.abc.MetaPathFinder):
         """Don't create a ``FSLeyesPluginFinder``. Instead, access the
         singleton instance via the :meth:`instance` method.
         """
-        FSLeyesPluginFinder._instance = self
         self.__plugins = {}
 
 
-    def add_plugin(self, module : ModuleType, modname : str, builtin : bool):
+    def add_plugin(self, module : ModuleType, modname : str):
         """Register a FSLeyes plugin module.
 
         :arg module:  The loaded module
         :arg modname: The module name
-        :arg builtin: Whether or not this is a built-in FSLeyes plugin, from
-                      the :mod:`fsleyes.plugins` module.
         """
-        self.__plugins[modname] = FSLeyesPlugin(module, modname, builtin)
+        self.__plugins[modname] = FSLeyesPlugin(module, modname)
 
 
     def find_distributions(self, context=None):
@@ -406,14 +488,29 @@ def initialise():
             log.warning('Failed to load plugin file %s: %s', fname, e)
 
 
-def _pluginGroup(cls : Plugin) -> Optional[str]:
-    """Returns the type/group of the given plugin, one of ``'views'``,
-    ``'controls'``, or ``'tools'``.
+def _pluginType(item) -> Union[str, bool]:
+    """Return the plugin type of the given object - one of ``'view'``,
+    ``'control'``, ``'tool'`` or ``'layout'``.
     """
-    if   issubclass(cls, viewpanel.ViewPanel):      return 'views'
-    elif issubclass(cls, ctrlpanel.ControlPanel):   return 'controls'
-    elif issubclass(cls, ctrlpanel.ControlToolBar): return 'controls'
-    elif issubclass(cls, actions.Action):           return 'tools'
+
+    if isinstance(item, type):
+        if   issubclass(item, ctrlpanel.ControlMixin): return 'control'
+        elif issubclass(item, viewpanel.ViewPanel):    return 'view'
+        elif issubclass(item, actions.Action):         return 'tool'
+    elif isinstance(item, (str, tuple)):               return 'layout'
+
+    return False
+
+
+def _pluginGroup(plg : Plugin) -> Optional[str]:
+    """Returns the type/group of the given plugin, one of ``'views'``,
+    ``'controls'``, ``'tools'``, or ``'layouts'``.
+    """
+    ptype = _pluginType(plg)
+    if   ptype == 'view':    return 'views'
+    elif ptype == 'control': return 'controls'
+    elif ptype == 'tool':    return 'tools'
+    elif ptype == 'layout':  return 'layouts'
     return None
 
 
@@ -432,18 +529,35 @@ def _loadBuiltIns():
         for _, name, _ in submods:
             log.debug('Loading built-in plugin module %s', name)
             submod = importlib.import_module(name)
-            FSLeyesPluginFinder.instance().add_plugin(
-                submod, submod.__name__, True)
+            FSLeyesPluginFinder.instance().add_plugin(submod, submod.__name__)
 
 
-def _listEntryPoints(group : str) -> Dict[str, Plugin]:
-    """Returns a dictionary containing ``{name : type}`` entry points for the
+def _listEntryPoints(
+        group   : str,
+        showAll : bool = False,
+        load    : bool = True
+) -> Dict[str, Plugin]:
+    """Returns a dictionary containing ``{name : object}`` entry points for the
     given entry point group.
 
     https://docs.python.org/3/library/importlib.metadata.html#entry-points
+
+    :arg group:   One of ``'fsleyes_views'``, ``'fsleyes_controls``,
+                  ``'fsleyes_tools'``, or ``'fsleyes_layouts'``.
+
+    :arg showAll: If ``True``, all plugins, including from installed
+                  third-party packages will be included. Otherwise (the
+                  default) plugins from third-party packages which are not in
+                  :attr:`SHOW_THIRD_PARTY_PLUGINS` will be omitted.  :arg
+                  load:
+
+    :arg load:    If ``True`` (the default), the returned dictionary will
+                  contain loaded entry point objects. If ``False``, the entry
+                  points will not be loaded, and the returned dictionary will
+                  instead contain ``importlib.metadata.EntryPoint`` objects.
     """
 
-    items = collections.OrderedDict()
+    items = {}
     eps   = impmeta.entry_points()
 
     # Python >=3.10 returns an EntryPoints
@@ -452,10 +566,18 @@ def _listEntryPoints(group : str) -> Dict[str, Plugin]:
     except Exception: eps = eps.get(group, [])
 
     for ep in eps:
+        # filter out third-party plugins by module path
+        if not (showAll or ep.value.startswith('fsleyes.')):
+            if not shouldShowThirdPartyPlugin(ep.value.split(':')[0]):
+                log.debug('Filtering third party plugin: %s', ep.value)
+                continue
         if ep.name in items:
             log.debug('Overriding entry point %s [%s] with entry '
                       'point of the same name from', ep.name, group)
-        items[ep.name] = ep.load()
+
+        if load: items[ep.name] = ep.load()
+        else:    items[ep.name] = ep
+
     return items
 
 
@@ -485,6 +607,7 @@ def listControls(viewType : Optional[View] = None) -> Dict[str, Control]:
     ctrls = _listEntryPoints('fsleyes_controls')
 
     for name, cls in list(ctrls.items()):
+
         if not issubclass(cls, (ctrlpanel.ControlPanel,
                                 ctrlpanel.ControlToolBar)):
             log.debug('Ignoring fsleyes_controls entry point %s - '
@@ -501,7 +624,7 @@ def listControls(viewType : Optional[View] = None) -> Dict[str, Control]:
         # views that it supports, or only the specific
         # view classes returned by supportedViews?
         subclassok = True
-        if class_defines_method(cls, 'supportSubClasses'):
+        if 'supportSubClasses' in cls.__dict__:
             subclassok = cls.supportSubClasses()
 
         if viewType  is not None and \
@@ -532,9 +655,9 @@ def listTools(viewType : Optional[View] = None) -> Dict[str, Tool]:
             tools.pop(name)
             continue
 
-        supported = cls.supportedViews()
-
         if viewType is not None:
+
+            supported = cls.supportedViews()
             # If a viewType is provided, we don't
             # return view-independent views
             if (supported is None) or \
@@ -544,18 +667,42 @@ def listTools(viewType : Optional[View] = None) -> Dict[str, Tool]:
     return tools
 
 
-def _lookupPlugin(clsname : str, group : str) -> Optional[Plugin]:
-    """Looks up the FSLeyes plugin with the given class name. """
-    entries = _listEntryPoints(f'fsleyes_{group}')
-    for cls in entries.values():
-        if cls.__name__ == clsname:
-            return cls
+def listLayouts() -> Dict[str, Layout]:
+    """Returns a dictionary of ``{name : str}`` mappings containing
+    the custom layouts provided by all installed FSLeyes plugins.
+    """
+
+    layouts = _listEntryPoints('fsleyes_layouts', showAll=True)
+
+    for name, layout in list(layouts.items()):
+
+        if not isinstance(layout, (str, tuple)):
+            log.debug('Ignoring fsleyes_layouts entry point '
+                      '%s - not a string or tuple', name)
+            layouts.pop(name)
+            continue
+
+        # layout plugins may be either layout strings,
+        # or tuples containing (name, layout string)
+        if isinstance(layout, tuple):
+            layouts[name] = layout[1]
+
+    return layouts
+
+
+def _lookupPlugin(plgname : str, group : str) -> Optional[Plugin]:
+    """Looks up the FSLeyes plugin with the given name. """
+    entries = _listEntryPoints(f'fsleyes_{group}', True)
+    for name, plugin in entries.items():
+        if isinstance(plugin, (str, tuple)):
+            if isinstance(plugin, tuple):
+                if plugin[0] == plgname:
+                    return plugin[1]
+            elif name == plgname:
+                return plugin
+        elif plugin.__name__ == plgname:
+            return plugin
     return None
-
-
-def lookupView(clsName : str) -> View:
-    """Looks up the FSLeyes view with the given class name. """
-    return _lookupPlugin(clsName, 'views')
 
 
 def lookupControl(clsName : str) -> Control:
@@ -568,12 +715,23 @@ def lookupTool(clsName : str) -> Tool:
     return _lookupPlugin(clsName, 'tools')
 
 
+def layoutModule(name : str) -> str:
+    """Return the module that a given layout is defined iwthin. """
+    layouts = _listEntryPoints('fsleyes_layouts', showAll=True, load=False)
+
+    for ep in layouts.values():
+        if ep.name == name:
+            return ep.value.split(':')[0].split('.')[0]
+
+    raise ValueError(f'Could not find layout with name {name}')
+
+
 def pluginTitle(plugin : Plugin) -> Optional[str]:
     """Looks and returns up the title under which the given ``plugin`` is
     registered.
     """
     group   = _pluginGroup(plugin)
-    entries = _listEntryPoints(f'fsleyes_{group}')
+    entries = _listEntryPoints(f'fsleyes_{group}', True)
     for title, cls in entries.items():
         if cls is plugin:
             return title
@@ -582,8 +740,20 @@ def pluginTitle(plugin : Plugin) -> Optional[str]:
 def loadPlugin(filename : str):
     """Loads the given Python file as a FSLeyes plugin. """
 
-    modname = op.splitext(op.basename(filename))[0].strip('_')
-    modname = f'fsleyes_plugin_{modname}'
+    # Generate a unique module name for each file
+    def genModuleName():
+        salt    = ''.join(random.choices(string.ascii_letters, k=10))
+
+        # We locate single-file modules within the
+        # fsleyes.plugins package, so that the
+        # _listEntryPoints function can filter them.
+        modname = op.splitext(op.basename(filename))[0].strip('_')
+        modname = f'fsleyes.plugins.{modname}_{salt}'
+        return modname
+
+    modname = genModuleName()
+    while modname in sys.modules:
+        modname = genModuleName()
 
     log.debug('Importing %s as %s', filename, modname)
 
@@ -593,7 +763,7 @@ def loadPlugin(filename : str):
 
     sys.modules[modname] = module
 
-    FSLeyesPluginFinder.instance().add_plugin(module, modname, False)
+    FSLeyesPluginFinder.instance().add_plugin(module, modname)
 
 
 def installPlugin(filename : str):