+++ fsleyes/tests/testdata/fsleyes_plugin_bad_example/fsleyes_plugin_bad_example/plugin.py
@@ -8,6 +8,9 @@ import fsleyes.panel as fslpanel
 # None of the things below inherit from the approprate
 # base-class, so will not be detected as FSLeyes plugins.
 
+class PluginLayout:
+    pass
+
 class PluginView(fslpanel.FSLeyesPanel):
 
     def __init__(self, *args, **kwargs):